LayoutTests/imported/w3c/web-platform-tests/css/css-inline/text-box-trim/text-box-trim-om-001-expected.txt

A

PASS getBoundingClientRect.top
FAIL getBoundingClientRect.height assert_equals: expected 20 but got 200
FAIL elementFromPoint: 10px above the client rect assert_equals: expected Element node <div id="target">A</div> but got Element node <div class="spacer"></div>
PASS elementFromPoint: 10px below the client rect
